Fix Mobile PWA not showing in true Fullscreen (#394)
* Fixed Rounding Error For TargetHalfHpDamageAttr

Moves that deal half of a target's HP were not able to deal damage if the target had 1 HP. Used Math.max to ensure 1 is the lowest this value ever evaluates to.

* Fix Mobile PWA not showing in true Fullscreen

adding "fullscreen" to the display attribute wasn't enough, but adding it to the start_url attribute was. This does not appear to affect PC (which is good since it shouldn't).

* Allow any orientation for mobile PWA

remove the orientation restriction since it doesn't prevent fullscreen PWA from displaying properly.