[Community] Add Reddit link to menu UI (#2338)
* feat: Add Reddit link to menu UI

Added a new menu option for Reddit with a corresponding handler to open the Reddit page in a new tab. This enhances the user experience by providing easy access to the Pokerogue subreddit.

* feat: Update wiki URL in menu UI

The wiki URL in the menu UI has been updated to "https://wiki.pokerogue.net/start" to provide a more specific starting point for users.