[Beta] Add deployment workflow for beta environment (#2595)
* feat: Add deployment workflow for beta environment

This commit adds a new GitHub Actions workflow file, deploy-beta.yml, which is responsible for deploying the application to the beta environment. The workflow is triggered on push and pull request events. It checks if the repository is 'pagefaultgames/pokerogue' before running the deployment steps. The deployment includes checking out the code, setting up the Node.js environment, installing dependencies, building the application, setting up SSH for secure communication, deploying the build on the server using rsync, and purging the Cloudflare cache. The deployment is only performed when the push event is triggered on the default branch.
It also changes the deploy.yml workflow to be triggered only when a release is generated.

* feat: Update deployment workflow for beta environment

* Update deploy-beta name

* chore: Update deploy-beta.yml to use 'f-fsantos:beta-environment' as the ref name

* chore: Update deploy-beta.yml to use 'f-fsantos:beta-environment' as the ref name

* chore: Update deploy-beta.yml to use '2595/merge' as the ref name

* chore: Update deploy-beta.yml to include event name in deployment message

* chore: Update deploy-beta.yml to use 'pull_request' event and '2595/merge' as the ref name

* chore: Update deploy-beta.yml to include SSH public key

* chore: Update deploy-beta.yml to include SSH public key

---------

Co-authored-by: Temps Ray <temps.ray@gmail.com>